feture - remove search results when user return home by click to brand name
